Remove gneral matrix routines from tensor math library and put in own module
