+++ optimism/test/test_TensorMath.py
@@ -4,7 +4,6 @@ from scipy.spatial.transform import Rotation
 import jax
 from jax import numpy as np
 from jax.test_util import check_grads
-from jax.scipy import linalg
 
 from optimism.test.TestFixture import TestFixture
 from optimism import TensorMath
@@ -26,18 +25,10 @@ def numerical_grad(f):
     return lam
 
 
-def generate_n_random_symmetric_matrices(n, minval=0.0, maxval=1.0):
-    key = jax.random.PRNGKey(0)
-    As = jax.random.uniform(key, (n,3,3), minval=minval, maxval=maxval)
-    return jax.vmap(lambda A: np.dot(A.T,A), (0,))(As)
-
-
 class TensorMathFixture(TestFixture):
 
     def setUp(self):
         self.log_squared = lambda A: np.tensordot(TensorMath.log_sqrt(A), TensorMath.log_sqrt(A))
-        self.sqrtm_jit = jax.jit(TensorMath.sqrtm)
-        self.logm_iss_jit = jax.jit(TensorMath.logm_iss)
         
 
     def test_log_sqrt_tensor_jvp_0(self):
@@ -205,127 +196,7 @@ class TensorMathFixture(TestFixture):
         check_grads(pow_squared, (C,), order=1)
         
         
-    ### sqrtm ###
-
-    
-    def test_sqrtm_jit(self):
-        C = generate_n_random_symmetric_matrices(1)[0]
-        sqrtC = self.sqrtm_jit(C)
-        self.assertFalse(np.isnan(sqrtC).any())
-        
-
-    def test_sqrtm(self):
-        mats = generate_n_random_symmetric_matrices(100)
-        sqrtMats = jax.vmap(self.sqrtm_jit, (0,))(mats)
-        shouldBeMats = jax.vmap(lambda A: np.dot(A,A), (0,))(sqrtMats)
-        self.assertArrayNear(shouldBeMats, mats, 10)
-
-
-    def test_sqrtm_fwd_mode_derivative(self):
-        C = generate_n_random_symmetric_matrices(1)[0]
-        check_grads(TensorMath.sqrtm, (C,), order=2, modes=["fwd"])
-
-
-    def test_sqrtm_rev_mode_derivative(self):
-        C = generate_n_random_symmetric_matrices(1)[0]
-        check_grads(TensorMath.sqrtm, (C,), order=2, modes=["rev"])
-
-
-    def test_sqrtm_on_degenerate_eigenvalues(self):
-        C = R@np.diag(np.array([2., 0.5, 2]))@R.T
-        sqrtC = TensorMath.sqrtm(C)
-        shouldBeC = np.dot(sqrtC, sqrtC)
-        self.assertArrayNear(shouldBeC, C, 12)
-        check_grads(TensorMath.sqrtm, (C,), order=2, modes=["rev"])
-
-
-    def test_sqrtm_on_10x10(self):
-        key = jax.random.PRNGKey(0)
-        F = jax.random.uniform(key, (10,10), minval=1e-8, maxval=10.0)
-        C = F.T@F
-        sqrtC = TensorMath.sqrtm(C)
-        shouldBeC = np.dot(sqrtC,sqrtC)
-        self.assertArrayNear(shouldBeC, C, 11)
-
-
-    def test_sqrtm_derivatives_on_10x10(self):
-        key = jax.random.PRNGKey(0)
-        F = jax.random.uniform(key, (10,10), minval=1e-8, maxval=10.0)
-        C = F.T@F
-        check_grads(TensorMath.sqrtm, (C,), order=1, modes=["fwd", "rev"])
-
-
-    def test_logm_iss_on_matrix_near_identity(self):
-        key = jax.random.PRNGKey(0)
-        id_perturbation = 1.0 + jax.random.uniform(key, (3,), minval=1e-8, maxval=0.01)
-        A = np.diag(id_perturbation)
-        logA = TensorMath.logm_iss(A)
-        self.assertArrayNear(logA, np.diag(np.log(id_perturbation)), 12)
-
-
-    def test_logm_iss_on_double_degenerate_eigenvalues(self):
-        eigvals = np.array([2., 0.5, 2.])
-        C = R@np.diag(eigvals)@R.T
-        logC = TensorMath.logm_iss(C)
-        logCSpectral = R@np.diag(np.log(eigvals))@R.T
-        self.assertArrayNear(logC, logCSpectral, 12)
-
-
-    def test_logm_iss_on_triple_degenerate_eigvalues(self):
-        A = 4.0*np.identity(3)
-        logA = TensorMath.logm_iss(A)
-        self.assertArrayNear(logA, np.log(4.0)*np.identity(3), 12)
-
-
-    def test_logm_iss_jit(self):
-        C = generate_n_random_symmetric_matrices(1)[0]
-        logC = self.logm_iss_jit(C)
-        self.assertFalse(np.isnan(logC).any())
-
-
-    def test_logm_iss_on_full_3x3s(self):
-        mats = generate_n_random_symmetric_matrices(1000)
-        logMats = jax.vmap(self.logm_iss_jit, (0,))(mats)
-        shouldBeMats = jax.vmap(lambda A: linalg.expm(A), (0,))(logMats)
-        self.assertArrayNear(shouldBeMats, mats, 7)      
-
-        
-    def test_logm_iss_fwd_mode_derivative(self):
-        C = generate_n_random_symmetric_matrices(1)[0]
-        check_grads(self.logm_iss_jit, (C,), order=1, modes=['fwd'])
-
-
-    def test_logm_iss_rev_mode_derivative(self):
-        C = generate_n_random_symmetric_matrices(1)[0]
-        check_grads(self.logm_iss_jit, (C,), order=1, modes=['rev'])
-
-
-    def test_logm_iss_hessian_on_double_degenerate_eigenvalues(self):
-        eigvals = np.array([2., 0.5, 2.])
-        C = R@np.diag(eigvals)@R.T
-        check_grads(jax.jacrev(TensorMath.logm_iss), (C,), order=1, modes=['fwd'], rtol=1e-9, atol=1e-9, eps=1e-5)
-
-
-    def test_logm_iss_derivatives_on_double_degenerate_eigenvalues(self):
-        eigvals = np.array([2., 0.5, 2.])
-        C = R@np.diag(eigvals)@R.T
-        check_grads(TensorMath.logm_iss, (C,), order=1, modes=['fwd'])
-        check_grads(TensorMath.logm_iss, (C,), order=1, modes=['rev'])
-
-
-    def test_logm_iss_derivatives_on_triple_degenerate_eigenvalues(self):
-        A = 4.0*np.identity(3)
-        check_grads(TensorMath.logm_iss, (A,), order=1, modes=['fwd'])
-        check_grads(TensorMath.logm_iss, (A,), order=1, modes=['rev'])
-
-
-    def test_logm_iss_on_10x10(self):
-        key = jax.random.PRNGKey(0)
-        F = jax.random.uniform(key, (10,10), minval=1e-8, maxval=10.0)
-        C = F.T@F
-        logC = TensorMath.logm_iss(C)
-        logCSpectral = TensorMath.logh(C)
-        self.assertArrayNear(logC, logCSpectral, 12)
+   
 
         
 if __name__ == '__main__':
